Issue #1 - Implemented various versions of StarSpace embeddings generator
