+++ categorical_variables_embeddings/generate_dense_datasets.py
@@ -1,68 +1,16 @@
 import logging
-import re
 import os
 
 import gensim
 import pandas as pd
 import numpy as np
-from abc import ABC, abstractmethod
 
 from modeling.utils import read_parquet
+from categorical_variables_embeddings.dense_dataset_generator import DenseDatasetGenerator
 
 # in order to use starspace, you need to install it using bin/install_starspace.sh
 # unfortunately, it won't be visible in an IDE
 import starwrap as sw
-
-class SentencesProcessorIterator:
-    def __init__(self, sentences):
-        self.regexp = re.compile('[\W_]+', re.UNICODE)
-        self.sentences = sentences
-
-    def __iter__(self):
-        for sentence in self.sentences:
-            yield self.regexp.sub(' ', sentence.lower()).split()
-
-
-class DenseDatasetGenerator(ABC):
-
-    def transform(self, dataset, id_attr, sentence_attr):
-        # fit_transform would indicate it's always trained on the dataset given which is not always the case
-        """
-        May perform fitting during the transformation.
-        """
-        processed_sentences = self._extract_processed_sentences(dataset, sentence_attr)
-        vectors_container = self._get_vectors_container(processed_sentences)
-        mean_vec = self._get_mean_vector(vectors_container)
-
-        sentence_vectors = []
-        for words in processed_sentences:
-            word_vecs = []
-            for w in words:
-                if w in vectors_container:
-                    word_vecs.append(vectors_container[w])
-
-            if len(word_vecs) > 0:
-                sentence_vectors.append(np.array(word_vecs).mean(axis=0))
-            else:
-                sentence_vectors.append(mean_vec)
-
-        sentence_vectors = pd.DataFrame(np.array(sentence_vectors))
-        sentence_vectors.columns = ["{}_{}".format(id_attr, c) for c in sentence_vectors.columns]
-        sentence_vectors[id_attr] = dataset[id_attr].values
-
-        return sentence_vectors
-
-    @abstractmethod
-    def _get_vectors_container(self, processed_sentences):
-        pass
-
-    def _get_mean_vector(self, vectors_container):
-        return np.array(list(vectors_container.values()), dtype='float32').mean(axis=0)
-
-    @staticmethod
-    def _extract_processed_sentences(dataset, sentence_attr):
-        raw_sentences = dataset[sentence_attr].tolist()
-        return SentencesProcessorIterator(raw_sentences)
 
 
 class GensimDenseGenerator(DenseDatasetGenerator):
@@ -129,7 +77,6 @@ class StarspaceDenseGenerator(DenseDatasetGenerator):
         arg.epoch = 20
         arg.dim = 50
         arg.similarity = "dot"
-        # arg.trainWord = 1???????????????????
         sp = sw.starSpace(arg)
         sp.init()
         sp.train()
@@ -172,7 +119,7 @@ class StarspaceDenseGenerator(DenseDatasetGenerator):
 
         sentence_vectors = pd.DataFrame(np.array(sentence_vectors))
         sentence_vectors.columns = ["{}_{}".format(id_attr, c) for c in sentence_vectors.columns]
-        sentence_vectors[id_attr] = dataset[id_attr].values
+        sentence_vectors[id_attr] = dataset[id_attr].to_numpy()
 
         os.remove("tmp_sentences.txt")
         os.remove("tmp_embeds.tsv")
@@ -281,9 +228,9 @@ if __name__ == "__main__":
     # *** Choose your generator
     # generator = GensimDenseGenerator()
     # generator = FileBasedDenseGenerator("data/wiki.ru.vec")
-    # generator = StarspaceDenseGenerator()
-    generator = StarspaceDenseGeneratorLabelDoc("data/items.csv", "data/item_categories.csv", "data/shops.csv",
-                                                "data/processed_full/train.parquet", "data/processed_full/test.parquet")
+    generator = StarspaceDenseGenerator()
+    # generator = StarspaceDenseGeneratorLabelDoc("data/items.csv", "data/item_categories.csv", "data/shops.csv",
+    #                                             "data/processed_full/train.parquet", "data/processed_full/test.parquet")
     # ***
 
     produce_dense_dataset(generator, "data/items.csv", 'item_id', 'item_name', "data/starspace_items_dense.parquet")
